Update thirdchance
In this case, I'm trying to develop and combine two algorithms to have only one algorithm that can search the name of a gene and the organism where it had been discovered and match this information with all the results that PDB can consider. This is the first part of this algorithm and one principal change I can make is to fix the XPath part, cause it's not working as I need.  Before, I tried using CSS, but it didn't work too. I suggest modifying the XPath part because it doesn't fit as it has to. If it doesn't work, maybe I have to look for some other alternatives that can show this information from the HTML code of genebank into this code. Also, I'm trying to develop the second algorithm but I have some modifications that I need to make about it because it is just a combination of some examples that I find on the internet and I'm not sure it works.